split docs-sync and cn-core roots
