+++ crates/cn-core/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod config;
+mod normalize;
