+++ crates/docs-sync/src/tests/mod.rs
@@ -1,0 +1,4 @@
+mod access;
+mod iroh_sync;
+mod memory;
+mod relay;
